LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-3.2-navigator-service-worker-expected.txt

CONSOLE MESSAGE: line 498: Partial interface Navigator with no original interface

FAIL Service Workers: navigator.serviceWorker Partial interface Navigator with no original interface
